# delivery/OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ResearchNavigator API",
    "description": "Evidence-grounded research direction and paper workspace.",
    "version": "2.2.2"
  },
  "paths": {
...
    "/api/evaluations/studies": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List Evaluation Studies",
        "operationId": "list_evaluation_studies_api_evaluations_studies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationStudyRead"
                  },
                  "type": "array",
                  "title": "Response List Evaluation Studies Api Evaluations Studies Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Create Evaluation Study",
        "operationId": "create_evaluation_study_api_evaluations_studies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationStudyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationStudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/evaluations/studies/{study_id}": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation Study",
        "operationId": "get_evaluation_study_api_evaluations_studies__study_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationStudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/studies/{study_id}/freeze": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Freeze Evaluation Study",
        "operationId": "freeze_evaluation_study_api_evaluations_studies__study_id__freeze_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationStudyRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/studies/{study_id}/assignments": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Create Evaluation Assignments",
        "operationId": "create_evaluation_assignments_api_evaluations_studies__study_id__assignments_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationAssignmentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationAssignmentRead"
                  },
                  "title": "Response Create Evaluation Assignments Api Evaluations Studies  Study Id  Assignments Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/assignments": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "List My Evaluation Assignments",
        "operationId": "list_my_evaluation_assignments_api_evaluations_assignments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationAssignmentRead"
                  },
                  "type": "array",
                  "title": "Response List My Evaluation Assignments Api Evaluations Assignments Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/evaluations/assignments/{assignment_id}/start": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Start Evaluation Assignment",
        "operationId": "start_evaluation_assignment_api_evaluations_assignments__assignment_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationAssignmentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/assignments/{assignment_id}/ratings": {
      "post": {
        "tags": [
          "evaluations"
        ],
        "summary": "Submit Evaluation Rating",
        "operationId": "submit_evaluation_rating_api_evaluations_assignments__assignment_id__ratings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRatingCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRatingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evaluations/studies/{study_id}/results": {
      "get": {
        "tags": [
          "evaluations"
        ],
        "summary": "Get Evaluation Results",
        "operationId": "get_evaluation_results_api_evaluations_studies__study_id__results_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Study Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResultRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationAssignmentCreate": {
        "properties": {
          "expert_user_id": {
            "type": "integer",
            "title": "Expert User Id"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated",
            "default": false
          },
          "task_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Ids"
          }
        },
        "type": "object",
        "required": [
          "expert_user_id"
        ],
        "title": "EvaluationAssignmentCreate"
      },
      "EvaluationAssignmentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_simulated": {
            "type": "boolean",
            "title": "Is Simulated"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/EvaluationVariantRead"
            },
            "type": "array",
            "title": "Variants"
          },
          "assigned_at": {
            "type": "string",
            "format": "date-time",
            "title": "Assigned At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "task_id",
          "task_key",
          "status",
          "is_simulated",
          "variants",
          "assigned_at",
          "started_at",
          "completed_at"
        ],
        "title": "EvaluationAssignmentRead"
      },
      "EvaluationRatingCreate": {
        "properties": {
          "evidence_correctness": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Evidence Correctness"
          },
          "evidence_sufficiency": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Evidence Sufficiency"
          },
          "citation_usefulness": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Citation Usefulness"
          },
          "missing_field_correctness": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Missing Field Correctness"
          },
          "preference": {
            "type": "string",
            "enum": [
              "A",
              "B",
              "tie"
            ],
            "title": "Preference"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "type": "object",
        "required": [
          "evidence_correctness",
          "evidence_sufficiency",
          "citation_usefulness",
          "missing_field_correctness",
          "preference"
        ],
        "title": "EvaluationRatingCreate"
      },
      "EvaluationRatingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "assignment_id": {
            "type": "integer",
            "title": "Assignment Id"
          },
          "preference": {
            "type": "string",
            "title": "Preference"
          },
          "duration_seconds": {
            "type": "integer",
            "title": "Duration Seconds"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "assignment_id",
          "preference",
          "duration_seconds",
          "submitted_at"
        ],
        "title": "EvaluationRatingRead"
      },
      "EvaluationResultRead": {
        "properties": {
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "real_expert_count": {
            "type": "integer",
            "title": "Real Expert Count"
          },
          "simulated_count": {
            "type": "integer",
            "title": "Simulated Count"
          },
          "validation_status": {
            "type": "string",
            "title": "Validation Status"
          },
          "claim_boundary": {
            "type": "string",
            "title": "Claim Boundary"
          }
        },
        "type": "object",
        "required": [
          "study_id",
          "metrics",
          "real_expert_count",
          "simulated_count",
          "validation_status",
          "claim_boundary"
        ],
        "title": "EvaluationResultRead"
      },
      "EvaluationStudyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "study_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Study Version"
          },
          "randomized_seed": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Randomized Seed"
          },
          "protocol": {
            "additionalProperties": true,
            "type": "object",
            "title": "Protocol"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/EvaluationTaskCreate"
            },
            "type": "array",
            "maxItems": 500,
            "minItems": 1,
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "name",
          "study_version",
          "randomized_seed",
          "tasks"
        ],
        "title": "EvaluationStudyCreate"
      },
      "EvaluationStudyRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "study_version": {
            "type": "string",
            "title": "Study Version"
          },
          "frozen_input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frozen Input Hash"
          },
          "expert_outcome_validation": {
            "type": "string",
            "title": "Expert Outcome Validation"
          },
          "randomized_seed": {
            "type": "string",
            "title": "Randomized Seed"
          },
          "protocol": {
            "additionalProperties": true,
            "type": "object",
            "title": "Protocol"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/EvaluationTaskRead"
            },
            "type": "array",
            "title": "Tasks"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "status",
          "study_version",
          "frozen_input_hash",
          "expert_outcome_validation",
          "randomized_seed",
          "protocol"
        ],
        "title": "EvaluationStudyRead"
      },
      "EvaluationTaskCreate": {
        "properties": {
          "task_key": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Task Key"
          },
          "paper_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Id"
          },
          "baseline_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Baseline Payload"
          },
          "candidate_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Candidate Payload"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "task_key",
          "baseline_payload",
          "candidate_payload",
          "position"
        ],
        "title": "EvaluationTaskCreate"
      },
      "EvaluationTaskRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "task_key": {
            "type": "string",
            "title": "Task Key"
          },
          "paper_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paper Id"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_key",
          "paper_id",
          "position"
        ],
        "title": "EvaluationTaskRead"
      },
      "EvaluationVariantRead": {
        "properties": {
          "label": {
            "type": "string",
            "enum": [
              "A",
              "B"
            ],
            "title": "Label"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "label",
          "payload"
        ],
        "title": "EvaluationVariantRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}